refactor(store): module simplification files
