+++ packages/store/tests/action.spec.ts
@@ -7,7 +7,7 @@ import { Action } from '../src/decorators/action';
 import { State } from '../src/decorators/state';
 import { META_KEY } from '../src/symbols';
 
-import { NgxsModule } from '../src/module';
+import { NgxsModule } from '../src/ngxs.module';
 import { Store } from '../src/store';
 import { Actions } from '../src/actions-stream';
 import {
